requirements.txt
beautifulsoup4==4.6.3
discord.py[voice]
pytz==2019.3
requests==2.21.0